Refatoração e correção de problemas no projeto EcoGuardians
