+++ Procfile
@@ -1,2 +1,2 @@
-web: gunicorn --worker-class eventlet -w 1 app:create_app() -b 0.0.0.0:$PORT
+web: gunicorn --worker-class eventlet -w 1 app:app -b 0.0.0.0:$PORT
